requirements.txt
chardet==5.2.0
faiss_cpu==1.12.0
numpy==2.2.6
openvino_genai==2025.3.0.0
pandas==2.3.3
optimum-intel[openvino]==1.26.1